Sistem_2_blockchain/Owner/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.7.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.5
click==8.1.8
cryptography==46.0.2
cytoolz==1.0.1
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
hexbytes==1.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
propcache==0.4.0
py-solc-x==2.0.4
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
pyunormalize==17.0.0
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
SQLAlchemy==2.0.43
toolz==1.0.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
web3==7.13.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
